Resume downloads after app close
it'll re-subscribe and everything!

took a few tries to get right but only bc of dumb debugging mistakes.
the code was really easy actually. biggest hurdle was that the ID
doesn't indicate the course directory, which is a thing I just added bc
it makes way more sense this way (easier to delete all of a course's
data). hmm just realized the metadata is still in the root folder, that
commit is coming in a few minutes