Remove metrics tracking for "Report a Problem"
this is a privacy issue, since it could help us link user activity to
email addresses. if we ever add problem reporting with a more fancy
mechanism than email, we could make it anonymous, but for now we need to
avoid logging this.

To compensate, we're auto-inserting the current track position into the
email, which should help track down the important stuff.