Use native queue
this should fix issues on devices that insist on adding next/prev
buttons
I think it'll also help pre-stream future tracks
it also makes for a nicer notification experience
BUT it may be a bit more buggy. we'll seeeeeee

I did have to nix a couple autoplay-related settings. "autoplay only
downloaded tracks" may be possible -- best way to do it is just to only
enqueue contiguous downloaded tracks -- but I don't think it's really
worth (what if the current track isn't downloaded? what if the user
wants to skip to the next track?). we could also just stop playback if
it autoplays to a non-downloaded track, but that has its own issues (and
the audio player may pre-buffer anyway)