Remove bundled lessons in Android
this is a refactor that avoids special-casing the first lesson, *except*
that when the app first loads, the tracks are copied from the bundle to
the normal storage space. This is indeed a little less efficient (stores
the tracks in the app bundle and in the app storage), but I think it's
worth it to avoid muddying the code with this iOS-only requirement (from
the App Store; they told us we need to bundle some content with the
app).

The pseudo-downloads are quick enough that I don't mind just firing them
off when the app first starts and hoping they're done by the time the
user needs them. Deletion and redownloading will grab from the server in
case there are updates.

Added a babel transform to yeet the require() statements to be sure
these don't get included in the Android bundle. I've tested this to the
extent that I can without an iOS device, but should work.